% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bland_Altman.R
\name{Bland_Atlman}
\alias{Bland_Atlman}
\title{Function to display a bland plot in order to visually assess the agreement between CytOpt estimation
of the class proportions and the estimate of the class proportions provided through manual gating.}
\usage{
Bland_Atlman(proportions, additional_info_shape = NULL)
}
\arguments{
\item{proportions}{\code{data.frame} of true and proportion estimates from \code{CytOpt()}}

\item{additional_info_shape}{vector of additional information to be used for shape in the plot. Not implemented yet.}
}
\description{
Function to display a bland plot in order to visually assess the agreement between CytOpt estimation
of the class proportions and the estimate of the class proportions provided through manual gating.
}
