% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytopt_minmax.R
\name{cytopt_minmax_r}
\alias{cytopt_minmax_r}
\title{Function to estimate the type cell proportions in an unclassified cytometry data set denoted X_s
by using the classification Lab_source from an other cytometry data set X_s. With this function
an additional regularization parameter on the class proportions enables a faster computation of the
estimator.}
\usage{
cytopt_minmax_r(
  X_s,
  X_t,
  Lab_source,
  theta_true = theta_true,
  eps = 1e-04,
  lbd = 1e-04,
  n_iter = 4000,
  step = 5,
  power = 0.99,
  monitoring = T
)
}
\arguments{
\item{X_s}{Cytometry data set. The columns correspond to the different biological markers tracked.
One line corresponds to the cytometry measurements performed on one cell. The classification
of this Cytometry data set must be provided with the Lab_source parameters.}

\item{X_t}{Cytometry data set. The columns correspond to the different biological markers tracked.
One line corresponds to the cytometry measurements performed on one cell. The CytOpT algorithm
targets the cell type proportion in this Cytometry data set.}

\item{Lab_source}{Classification of the X_s Cytometry data set}

\item{theta_true}{True proportions in the target data set derived from manual gating.}

\item{eps}{Regularization parameter of the Wasserstein distance}

\item{step}{Constant that multiply the step-size policy}

\item{power}{the step size policy of the gradient ascent method is step/n^power.}

\item{monitoring}{a logical flag indicating to possibly monitor the gap between the estimated proprotions and the manual
gold-standard. Default is \code{FALSE}}
}
\value{
A list with the following elements:\code{Results_Minmax}
}
\description{
Function to estimate the type cell proportions in an unclassified cytometry data set denoted X_s
by using the classification Lab_source from an other cytometry data set X_s. With this function
an additional regularization parameter on the class proportions enables a faster computation of the
estimator.
}
\examples{
Stanford1A_values <- read.csv('tests/ressources/W2_1_values.csv')
Stanford1A_clust <- read.csv('tests/ressources/W2_1_clust.csv')[, 2]
Stanford3A_values <- read.csv('tests/ressources/W2_7_values.csv')
Stanford3A_clust <- read.csv('tests/ressources/W2_7_clust.csv')[, 2]
X_source <- convertArray(Stanford1A_values)
X_target <- convertArray(Stanford3A_values)
X_source <- X_source * (X_source > 0)

theta_true <- rep(0,10)
for (k in 1:10) theta_true[k] <- sum(Lab_target == k)/length(Lab_target)
cytopt_minmax_r(X_s=X_source, X_t=X_target,
               Lab_source=convertArray(Lab_source),
               theta_true=theta_true)
}
