% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL_plot.R
\name{KL_plot}
\alias{KL_plot}
\title{Kullback-Leibler divergence plot}
\usage{
KL_plot(
  monitoring,
  n_0 = 10,
  n_stop = 1000,
  title = "Kullback-Liebler divergence trace"
)
}
\arguments{
\item{monitoring}{\code{list} of monitoring estimates from \code{CytOpt()} output.}

\item{n_0}{first iteration to plot. Default is 10.}

\item{n_stop}{last iteration to plot. Default is 1000.}

\item{title}{plot title. Default is \code{"Kullback-Liebler divergence trace"}.}
}
\description{
A plotting function for displaying Kullback-Liebler (KL) divergence across 
iterations of the optimization algorithm(s).
}
