% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab_source.R
\docType{data}
\name{Lab_source}
\alias{Lab_source}
\title{Lab_source}
\format{
A with 1 column, with are:
\describe{
  \item{x}{label source}
}
}
\usage{
Lab_source
}
\description{
A simple that is used as an example for the analysis test extracted from the Stanfort data set.
}
\keyword{datasets}
