% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab_target.R
\docType{data}
\name{Lab_target}
\alias{Lab_target}
\title{Lab_target}
\format{
A with 1 column, with are:
\describe{
  \item{x}{label}
}
}
\usage{
Lab_target
}
\description{
A simple that is used as an example for the analysis test extracted from the Stanfort data set.
}
\keyword{datasets}
