% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Label_Prop_sto.R
\name{Label_Prop_sto_r}
\alias{Label_Prop_sto_r}
\title{Function to.}
\usage{
Label_Prop_sto_r(X_s, X_t, levels, h_hat, eps = 1e-04, n_iter = 4000)
}
\arguments{
\item{X_s}{a cytometry dataframe. The columns correspond to the different biological markers tracked.
One line corresponds to the cytometry measurements performed on one cell. The classification
of this Cytometry data set must be provided with the Lab_source parameters.}

\item{X_t}{a cytometry dataframe. The columns correspond to the different biological markers tracked.
One line corresponds to the cytometry measurements performed on one cell. The CytOpT algorithm
targets the cell type proportion in this Cytometry data set}

\item{levels}{a vector of the classe}

\item{h_hat}{a vector of the estimations obtained method select}

\item{eps}{an float value of regularization parameter of the Wasserstein distance. Default is \code{1e-04}}

\item{n_iter}{an integer Constant that iterate method select. Default is \code{4000}}
}
\value{
A list with the following elements:\itemize{
  \item \code{Lab_target_hat_one}
  \item \code{Lab_target_hat_two}
}
}
\description{
Function to.
}
