% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miami3A.R
\docType{data}
\name{Miami3A}
\alias{Miami3A}
\title{Miami3A

Here we create a list Miami3A, of two components with data types of type dataframe.}
\format{
The Miami3A data list is divided into two elements:
\describe{
  \item{values}{The values measured}
  \item{clust}{The clust of population}
}
}
\usage{
data(Miami3A)
}
\description{
Miami3A

Here we create a list Miami3A, of two components with data types of type dataframe.
}
\keyword{data}
