% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miami3A_Target.R
\docType{data}
\name{Miami3A_clust}
\alias{Miami3A_clust}
\title{Miami3A_clust

Here we create a dataframe with Miami data.}
\format{
An object of class \code{integer} of length 42394.
}
\usage{
data(Miami3A_clust)
}
\description{
Miami3A_clust

Here we create a dataframe with Miami data.
}
\keyword{data}
