% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miami3A_source.R
\docType{data}
\name{Miami3A_values}
\alias{Miami3A_values}
\title{Miami3A_values

Here we create a dataframe with Miami data.}
\format{
An object of class \code{data.frame} with 42394 rows and 7 columns.
}
\usage{
data(Miami3A_values)
}
\description{
Miami3A_values

Here we create a dataframe with Miami data.
}
\keyword{data}
