% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.CytOpt.R
\name{print.CytOpt}
\alias{print.CytOpt}
\title{CytOpt print}
\usage{
\method{print}{CytOpt}(x, ...)
}
\arguments{
\item{x}{an object of class \code{CytOpt} a summary is desired.}

\item{...}{further arguments passed to or from other methods. Not implemented.}
}
\description{
print S3 method for CytOpt object
}
