% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoadPython.R
\name{run_python_files}
\alias{run_python_files}
\title{Function to Calling Python from R with reticulate package}
\usage{
run_python_files(
  py_files = c("/home/kalidou/PycharmProjects/CytOpT.R/CytOpT_pkg/CytOpt_plot.py",
    "/home/kalidou/PycharmProjects/CytOpT.R/CytOpT_pkg/minMaxScale.py",
    "/home/kalidou/PycharmProjects/CytOpT.R/CytOpT_pkg/Tools_CytOpt_Descent_Ascent.py",
    "/home/kalidou/PycharmProjects/CytOpT.R/CytOpT_pkg/Tools_CytOpt_MinMax_Swapping.py")
)
}
\arguments{
\item{py_files}{set of files python}
}
\description{
Function to Calling Python from R with reticulate package
}
