% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stanford3C.R
\docType{data}
\name{Stanford3C}
\alias{Stanford3C}
\title{Stanford3C

Here we create a list Stanford3C, of two components with data types of type dataframe.}
\format{
The Stanford3C data list is divided into two elements:
\describe{
  \item{values}{The values measured}
  \item{clust}{The clust of population}
}
}
\usage{
data(Stanford3C)
}
\description{
Stanford3C

Here we create a list Stanford3C, of two components with data types of type dataframe.
}
\keyword{data}
