% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stanford3C_Target.R
\docType{data}
\name{Stanford3C_clust}
\alias{Stanford3C_clust}
\title{Stanford3C_clust

Here we create a dataframe with Miami data.}
\format{
An object of class \code{integer} of length 33960.
}
\usage{
data(Stanford3C_clust)
}
\description{
Stanford3C_clust

Here we create a dataframe with Miami data.
}
\keyword{data}
