% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stanford3C_source.R
\docType{data}
\name{Stanford3C_values}
\alias{Stanford3C_values}
\title{Stanford3C_values

Here we create a dataframe with Miami data.}
\format{
An object of class \code{data.frame} with 33960 rows and 7 columns.
}
\usage{
data(Stanford3C_values)
}
\description{
Stanford3C_values

Here we create a dataframe with Miami data.
}
\keyword{data}
