% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.CytOpt.R
\name{summary.CytOpt}
\alias{summary.CytOpt}
\title{CytOpt summary}
\usage{
\method{summary}{CytOpt}(object, ...)
}
\arguments{
\item{object}{an object of class \code{CytOpt} for which a summary is desired.}

\item{...}{further arguments passed to or from other methods. Not implemented.}
}
\description{
summary S3 method for CytOpt object
}
