% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ucla2B.R
\docType{data}
\name{Ucla2B}
\alias{Ucla2B}
\title{Ucla2B

Here we create a list Ucla2B, of two components with data types of type dataframe.}
\format{
The Ucla2B data list is divided into two elements:
\describe{
  \item{values}{The values measured}
  \item{clust}{The clust of population}
}
}
\usage{
data(Ucla2B)
}
\description{
Ucla2B

Here we create a list Ucla2B, of two components with data types of type dataframe.
}
\keyword{data}
