% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X_source.R
\docType{data}
\name{X_source}
\alias{X_source}
\title{X_source}
\format{
A with 2 column, with are:
\describe{
  \item{CD4}{CD4}
  \item{CD8}{CD8}
}
}
\usage{
X_source
}
\description{
A simple that is used as an example for the analysis test extracted from the Stanfort data set.
}
\keyword{datasets}
