% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bland_Alman.R
\name{Bland_Atlman}
\alias{Bland_Atlman}
\title{Function to display a bland plot in order to visually assess the agreement between cytopt estimation
of the class proportions and the estimate of the class proportions provided through manual gating.}
\usage{
Bland_Atlman(proportions, additional_info_shape = NULL)
}
\arguments{
\item{estimates}{prop estimate with Desasc Opt method}

\item{True_Prop}{The benchmark estimate for the cell type proportions. It is provieded by the manual.
gating}

\item{Lab_source}{a vector of length \code{n} Classification of the X_s cytometry data set}
}
\description{
Function to display a bland plot in order to visually assess the agreement between cytopt estimation
of the class proportions and the estimate of the class proportions provided through manual gating.
}
