% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bland_Alman.R
\name{Bland_Atlman_r}
\alias{Bland_Atlman_r}
\title{Function to display a bland plot in order to visually assess the agreement between cytopt estimation
of the class proportions and the estimate of the class proportions provided through manual gating.}
\usage{
Bland_Atlman_r(Desac_hat, Minmax_hat, True_Prop, Lab_source)
}
\arguments{
\item{Desac_hat}{prop estimate with Desasc Opt method}

\item{Minmax_hat}{prop estimate with Min max Opt method}

\item{True_Prop}{The benchmark estimate for the cell type proportions. It is provieded by the manual.
gating}

\item{Lab_source}{a vector of length \code{n} Classification of the X_s cytometry data set}
}
\description{
Function to display a bland plot in order to visually assess the agreement between cytopt estimation
of the class proportions and the estimate of the class proportions provided through manual gating.
}
