% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytOpT.R
\name{CytOpT}
\alias{CytOpT}
\title{Function to estimate the type cell proportions in an unclassified cytometry data set denoted X_s
by using the classification Lab_source from an other cytometry data set X_s. With this function
the computation of the estimate of the class proportions is done with a descent ascent or minmax
or two algorithms.}
\usage{
CytOpT(
  X_s = NULL,
  X_t = NULL,
  Lab_source = NULL,
  Lab_target = NULL,
  theta_true = NULL,
  cell_type = NULL,
  names_pop = NULL,
  method = c("cytopt_minmax", "cytopt_desasc", "comparison_opt"),
  eps = 1e-04,
  n_iter = 4000,
  power = 0.99,
  step_grad = 50,
  step = 5,
  lbd = 1e-04,
  n_out = 1000,
  n_stoc = 10,
  n_0 = 10,
  n_stop = 1000,
  minMaxScaler = T,
  monitoring = T,
  thresholding = T
)
}
\arguments{
\item{X_s}{a cytometry dataframe. The columns correspond to the different biological markers tracked.
One line corresponds to the cytometry measurements performed on one cell. The classification
of this Cytometry data set must be provided with the Lab_source parameters.}

\item{X_t}{a cytometry dataframe. The columns correspond to the different biological markers tracked.
One line corresponds to the cytometry measurements performed on one cell. The CytOpT algorithm
targets the cell type proportion in this Cytometry data set}

\item{Lab_source}{a vector of length \code{n} Classification of the X_s cytometry data set}

\item{Lab_target}{a vector of length \code{n} classification of the X_t cytometry data set}

\item{theta_true}{If available, the true proportions in the target data set X_s. It allows to assess
the gap between the estimate of our method and the estimate of the cell type proportions derived from
manual gating.}

\item{cell_type}{a vector indicate type of cell}

\item{names_pop}{//commentary}

\item{method}{a character string indicating which method to use to
compute the cytopt, either \code{'cytopt_minmax'}, \code{'cytopt_desasc'}
or  \code{'comparison_opt'} for Comparison two methods Desasc or Minmax.
Default is \code{'cytopt_minmax'} since it is the method used in the test.}

\item{eps}{an float value of regularization parameter of the Wasserstein distance. Default is \code{1e-04}}

\item{n_iter}{an integer Constant that iterate method select. Default is \code{4000}}

\item{power}{an float constant the step size policy of the gradient ascent method is step/n^power. Default is \code{0.99}}

\item{step_grad}{an integer number step size of the gradient descent algorithm of the outer loop.
Default is \code{50}}

\item{step}{an integer constant that multiply the step-size policy. Default is \code{5}}

\item{lbd}{an float constant that multiply the step-size policy. Default is \code{1e-04}}

\item{n_out}{an integer number of iterations in the outer loop. This loop corresponds to the gradient
descent algorithm to minimize the regularized Wasserstein distance between the source and
target data sets. Default is \code{1000}}

\item{n_stoc}{an integer number of iterations in the inner loop. This loop corresponds to the stochastic
algorithm that approximates a maximizer of the semi dual problem. Default is \code{10}}

\item{n_0}{an integer value. Default is \code{10}}

\item{n_stop}{an integer value. Default is \code{1000}}

\item{minMaxScaler}{a logical flag indicating to possibly Scaler}

\item{monitoring}{a logical flag indicating to possibly monitor the gap between the estimated proprotions and the manual
gold-standard. Default is \code{FALSE}}

\item{thresholding}{a logical flag.}
}
\value{
A list with the following elements:\itemize{
  \item \code{h_hat}
  \item \code{Res_df}
  \item \code{Dico_res}
  \item \code{h_monitoring}
}
}
\description{
Function to estimate the type cell proportions in an unclassified cytometry data set denoted X_s
by using the classification Lab_source from an other cytometry data set X_s. With this function
the computation of the estimate of the class proportions is done with a descent ascent or minmax
or two algorithms.
}
