% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytOpT.R
\name{CytOpT}
\alias{CytOpT}
\title{Function to estimate the type cell proportions in an unclassified cytometry data set denoted X_s
by using the classification Lab_source from an other cytometry data set X_s. With this function
the computation of the estimate of the class proportions is done with a descent ascent or minmax
or two algorithms.}
\usage{
CytOpT(
  X_s = NULL,
  X_t = NULL,
  Lab_source = NULL,
  Lab_target = NULL,
  theta_true = NULL,
  cell_type = NULL,
  names_pop = NULL,
  method = c("minmax", "desasc", "both"),
  eps = 1e-04,
  n_iter = 4000,
  power = 0.99,
  step_grad = 50,
  step = 5,
  lbd = 1e-04,
  n_out = 1000,
  n_stoc = 10,
  minMaxScaler = TRUE,
  monitoring = TRUE,
  thresholding = TRUE
)
}
\arguments{
\item{X_s}{a cytometry dataframe. The columns correspond to the different biological markers tracked.
One line corresponds to the cytometry measurements performed on one cell. The classification
of this Cytometry data set must be provided with the Lab_source parameters.}

\item{X_t}{a cytometry dataframe. The columns correspond to the different biological markers tracked.
One line corresponds to the cytometry measurements performed on one cell. The CytOpT algorithm
targets the cell type proportion in this Cytometry data set}

\item{Lab_source}{a vector of length \code{n} Classification of the X_s cytometry data set}

\item{Lab_target}{a vector of length \code{n} classification of the X_t cytometry data set}

\item{theta_true}{If available, the true proportions in the target data set X_s. It allows to assess
the gap between the estimate of our method and the estimate of the cell type proportions derived from
manual gating.}

\item{cell_type}{a vector indicate type of cell}

\item{names_pop}{//commentary}

\item{method}{a character string indicating which method to use to
compute the cytopt, either \code{'minmax'}, \code{'desasc'}
or  \code{'both'} for comparing both Min-max swapping and descent-ascent procedures.
Default is \code{'minmax'}.}

\item{eps}{an float value of regularization parameter of the Wasserstein distance. Default is \code{1e-04}}

\item{n_iter}{an integer Constant that iterate method select. Default is \code{4000}}

\item{power}{an float constant the step size policy of the gradient ascent method is step/n^power. Default is \code{0.99}}

\item{step_grad}{an integer number step size of the gradient descent algorithm of the outer loop.
Default is \code{50}}

\item{step}{an integer constant that multiply the step-size policy. Default is \code{5}}

\item{lbd}{an float constant that multiply the step-size policy. Default is \code{1e-04}}

\item{n_out}{an integer number of iterations in the outer loop. This loop corresponds to the gradient
descent algorithm to minimize the regularized Wasserstein distance between the source and
target data sets. Default is \code{1000}}

\item{n_stoc}{an integer number of iterations in the inner loop. This loop corresponds to the stochastic
algorithm that approximates a maximizer of the semi dual problem. Default is \code{10}}

\item{minMaxScaler}{a logical flag indicating to possibly Scaler}

\item{monitoring}{a logical flag indicating to possibly monitor the gap between the estimated proprotions and the manual
gold-standard. Default is \code{FALSE}}

\item{thresholding}{a logical flag.}
}
\value{
a object of class \code{CytOpt}, which is a list of two elements:\itemize{
  \item \code{proportions} a \code{data.frame} with the (optionally true and) 
  estimated proportions for each \code{method}
  \item \code{monitoring} a list of estimates over the optimization iterations
  for each \code{method} (listed within)
}
}
\description{
Function to estimate the type cell proportions in an unclassified cytometry data set denoted X_s
by using the classification Lab_source from an other cytometry data set X_s. With this function
the computation of the estimate of the class proportions is done with a descent ascent or minmax
or two algorithms.
}
\examples{
if(interactive()){
head(X_source)
head(X_target)
head(Lab_source)
head(Lab_target)

# create theta true variable
theta_true <- rep(0,10)
for (k in 1:10) theta_true[k] <- sum(Lab_target == k)/length(Lab_target)

# Minmax swapping procedure
# parameters setting for the second procedure
lbd <- 0.0001
eps_two <- 0.0001
n_iter <- 10000
step_size <- 5
power <- 0.99

# Run Minmax swapping

# Build the Minmax swapping procedure
res <- CytOpT(X_source,X_target,Lab_source,theta_true=theta_true,
eps = eps_two, lbd = lbd, n_iter = n_iter, step = step_size, power = power, method='minmax')

# Summary res CytOpT
summary(res)

# Kullback-Leibler divergence \+ proportions plot
plot(res)

# BA plot
Bland_Atlman(res$proportions)
}
}
