% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/St_target.R
\docType{data}
\name{Lab_target}
\alias{Lab_target}
\title{Lab_target}
\format{
An object of class \code{integer} of length 33992.
}
\usage{
data(Lab_target)
}
\description{
A simple data that is used as an example for the analysis test extracted from the Stanfort data set.
Flow cytometry was measured in 3 samples for each 3 patients (ID: 1228, 1349 and 1369) replicated in 7 centers
(NHLBI, Yale, UCLA, CIMR, Baylor, Stanford and Miami), i.e. 63 files in total. Those .fcs files were converted
into .csv available under the directory: “flow_cytometry_data”. Manual gating from the different centers is
available under the “manual_gating vs gating” directory. A specific manual gating name has been assigned for each
center (see table below for correspondence). Manual gating results (i.e. cell population partition) from each
center is available under the directory “manual_gating vs clustering”. For each sample of each patient, there is
a “_values” file containing flow cytometry data only for gated populations and a “_clust” file to know the cell
allocation in the different populations.
}
\keyword{data}
